+++ xmodule/modulestore/xml.py
@@ -25,7 +25,7 @@ from xblock.runtime import DictKeyValueStore
 from common.djangoapps.util.monitoring import monitor_import_failure
 from xmodule.error_block import ErrorBlock
 from xmodule.errortracker import exc_info_to_str, make_error_tracker
-from xmodule.mako_module import MakoDescriptorSystem
+from xmodule.mako_block import MakoDescriptorSystem
 from xmodule.modulestore import COURSE_ROOT, LIBRARY_ROOT, ModuleStoreEnum, ModuleStoreReadBase
 from xmodule.modulestore.xml_exporter import DEFAULT_CONTENT_FIELDS
 from xmodule.tabs import CourseTabList
