+++ xmodule/poll_module.py
@@ -20,7 +20,7 @@ from lxml import etree
 from xblock.core import XBlock
 from xblock.fields import Boolean, Dict, List, Scope, String  # lint-amnesty, pylint: disable=wrong-import-order
 from openedx.core.djangolib.markup import Text, HTML
-from xmodule.mako_module import MakoTemplateBlockBase
+from xmodule.mako_block import MakoTemplateBlockBase
 from xmodule.stringify import stringify_children
 from xmodule.util.xmodule_django import add_webpack_to_fragment
 from xmodule.x_module import (
